Realised I was putting context.ts code in route.ts. Copied this code from route to context, then deleted the contents of route.ts
