+++ src/trpc/context.ts
@@ -1,18 +1,16 @@
-// trpc/context.ts
-import { inferAsyncReturnType } from "@trpc/server"; // Correct import
-import { CreateNextContextOptions } from '@trpc/server/adapters/next';
-import { getSession } from "next-auth"; // Adjust if using a different library or version
-import { authOptions } from '../auth'; //Adjust import based on auth setup
+import { auth } from "@/app/auth"; //Bringing in function called auth from /app/auth
+import { type Session } from "@next-auth"; //Imports a session from next-auth, a library used for auth in Next.js
 
-export const createContext = async ({ req, res }: CreateNextContextOptions) => {
-  const session = await getSession(req, res. authOptions); // Ensure getSession is properly defined
+export type Context = { //Creates a custom type called Context - essentially an object schema
+  session: Session | null; //States that session can either be a Session object (if someone is logged in) or null if no one is logged in
+  userID?: string; //States that userID is a string and is optional (?)
+  role?: string; //States that the role is a string and is an optional property (due to ?)
+}; 
 
-  return {
-    session,
-    userID: session?.user?.id ?? null,
-    role: session?.user?.role ?? 'USER', //Default role if none is assigned 
-  };
-};
+export async function createContext(): Promise<Context> { //Defines async function, 'createContext', returns promise of the Context object
+  const session = await auth(); //Calls auth function and waits to return a result, the result is stored in the session variable
 
-// Use inferAsyncReturnType to define Context
-export type Context = inferAsyncReturnType<typeof createContext>;
+
+}
+
+export { handler as GET, handler as POST };
